requirements.txt
# Core ML
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0

# Data
Pillow>=9.0.0
requests>=2.28.0
python-dotenv>=1.0.0

# Evaluation & Plotting
scikit-learn>=1.2.0
matplotlib>=3.7.0
numpy>=1.24.0

# Dataset download
kaggle>=1.5.0
